Fix build --install failure in SDK top level (#126)
When running build.py --install from top level, do not add clashing packaging/install rules.
The original if condition here only applied when --package was selected,
not for --install.
What we want is just to check if we are running from component level,
and only in that case bother with the packaging rules.

Change-Id: Ic6a1e037fc615a219676a9f0be89c53071ce0563

Signed-off-by: Gergely Nagy <gergely.nagy@arm.com>